### 2016-10-04 #86 @whedon
[PRE REVIEW]: Mordecai: Full Text Geoparsing and Event Geocoding
**Submitting author:** @ahalterman (<a href="http://orcid.org/0000-0001-9716-9555">Andrew Lee Halterman</a>)
**Repository:** <a href="https://github.com/openeventdata/mordecai" target ="_blank">https://github.com/openeventdata/mordecai</a>
**Version:** v1.1.0
