### 2016-10-12 #86 @acabunoc
Hey @jsta any interest in reviewing this?

Also flagging @joeyklee, @auremoser, @riordan in case any of you are interested in becoming a reviewer for JOSS (it's fun!)
