### 2016-10-12 #86 @riordan
Totally! I've been watching the mordecai project from afar from some time. I'd be down to be a reviewer.

I stumbled across this like a year ago when they were they first folks to start combining Word2Vec with Elasticsearch.
