### 2016-10-12 #86 @acabunoc
Yay thanks @riordan 🎉 !

@arfon can let you know how to get on the official reviewers list. In the meantime, you can read the [JOSS reviewer guidelines](http://joss.theoj.org/about#reviewer_guidelines) and I'll set you as reviewer here!
