### 2016-11-01 #91 @ahalterman
@riordan, let me know if you run into any trouble with the documentation/installation. The model's pretty big, unfortunately, so getting it stood up doesn't go smoothly depending on how much memory you have.
