### 2016-11-27 6c5dd65
Add support for models loaded from docker volumes
