+++ resources/utilities.py
@@ -2,74 +2,189 @@
 from __future__ import unicode_literals
 import os
 import sys
-import glob
 import json
+import numpy
 import mitie
+import pprint
+import argparse
+import unidecode
+from gensim.models import Word2Vec
 from elasticsearch_dsl import Search
 from ConfigParser import ConfigParser
 from elasticsearch import Elasticsearch
 from elasticsearch_dsl.query import MultiMatch
 
-# read in config file
-__location__ = os.path.realpath(os.path.join(os.getcwd(),
-                                             os.path.dirname(__file__)))
-config_file = glob.glob(os.path.join(__location__, '../config.ini'))
-parser = ConfigParser()
-parser.read(config_file)
-mitie_directory = parser.get('Locations', 'mitie_directory')
-mitie_ner_model = parser.get('Locations', 'mitie_ner_model')
+
+def parse_args():
+    parser = argparse.ArgumentParser(description='Mordecai Geolocation')
+    parser._optionals.title = 'Options'
+    parser.add_argument('-c', '--config-file',
+                        help='Specify path to config file.',
+                        type=str,
+                        required=False,
+                        default="")
+    parser.add_argument('-p', '--port',
+                        help='Specify port to listen on.',
+                        type=int,
+                        required=False,
+                        default=5000)
+    parser.add_argument('-eh', '--elasticsearch-host',
+                        help='Specify elasticsearch host.',
+                        type=str,
+                        required=False,
+                        default='elastic')
+    parser.add_argument('-ep', '--elasticsearch-port',
+                        help='Specify elasticsearch port.',
+                        type=str,
+                        required=False,
+                        default='9200')
+    parser.add_argument('-w', '--w2v-model',
+                        help='Specify path to w2v model.',
+                        type=str,
+                        required=False,
+                        default="/usr/src/data/GoogleNews-vectors-negative300.bin.gz")
+    parser.add_argument('-md', '--mitie-dir',
+                        help='Specify MITIE directory.',
+                        type=str,
+                        required=False,
+                        default="/usr/src/MITIE/mitielib")
+    parser.add_argument('-mn', '--mitie-ner',
+                        help='Specify path to MITIE NER model.',
+                        type=str,
+                        required=False,
+                        default="/usr/src/data/MITIE-models/english/ner_model.dat")
+    return parser.parse_args()
 
 
-def setup_mitie(mitie_directory):
-    """ Given the location for MITIE and the model, create a named_entity_extractor object."""
-    sys.path.append(mitie_directory)
-    ner_model = mitie.named_entity_extractor(mitie_ner_model)
-    return ner_model
-
-def setup_es():
-    """ 
-    Read the config file for where to find the geonames elasticsearch index.
+def get_configs(args):
+    '''
+    Given the command line arguments, first check for a config file. If there
+    is a config file use that, if ther eis no config file first check for
+    environment variables, then fall back to command line arguments.
 
     If geonames/ES is running on a different server, the Server section in the config
     file should be uncommented and filled in. If it's running locally and being linked
     through Docker (e.g. `sudo docker run -d -p 5000:5000 --link elastic:elastic mordecai`),
-    comment out the Server section so it knows to look for a linked container called `elastic` 
+    comment out the Server section so it knows to look for a linked container called `elastic`
     running on port 9200.
+    '''
+    config_dict = {}
+
+    # if there's a config file provided just use that
+    if args.config_file:
[…]
