### 2016-11-27 d8e76cc
Fix stopword_country_names.json import
