### 2016-11-28 fd57e65
Fix bugs
