### 2016-12-04 #91 @ahalterman
I merged a PR in that reorganizes how it deals with the data download, which is something you had changed, @riordan. This is sort of a general JOSS question, should I re-release the repo and change which version is being submitted? Or should we keep going with the existing version? The differences aren't too big so it shouldn't matter too much either way.
