### 2016-12-04 #91 @arfon
> This is sort of a general JOSS question, should I re-release the repo and change which version is being submitted? Or should we keep going with the existing version? The differences aren't too big so it shouldn't matter too much either way.

@ahalterman - I would recommend making a new release once the review is complete and all of the changes (associated with the review) have been made.
