### 2017-01-03 #91 @riordan
@ahalterman: Thank you for the prod! I finished my tests up 2 weeks ago and it's really impressive.

The installation ran according to documentation, which was really straightforward for installation. I'd suggest a clearer statement of need in the paper (something you do well in the README.md). Bringing in a bit more context, especially from your experience scanning news articles to find the places that are being written about, would be really useful here.

Other than that, I think it just needs an archival DOI (though I'd defer to @acabunoc or @arfon on that).
