### 2017-01-08 #91 @ahalterman
@riordan Thanks so much for doing this review! I'm glad you liked it and that things checked out. I just updated the paper to include more of the statement of need. If it looks good to you, it sounds like the paper ready to go. Thanks everyone!
