### 2017-01-18 #91 @ahalterman
I wanted to check in again and see if that commit addressed your last request, @riordan. If so, I think it should be ready for a DOI.
