### 2017-01-18 #91 @acabunoc
Thanks for the great work here @riordan & @ahalterman 👏  🎉

@ahalterman, could you make sure there's an archive of the reviewed software in Zenodo/figshare/other service and update this thread with the DOI of the archive? Thanks!
