### 2017-01-20 #91 @arfon
@ahalterman your paper is now accepted into JOSS and your paper DOI is http://dx.doi.org/10.21105/joss.00091 ⚡️ 🚀 💥
